requirements.txt
numpy==1.23.3
matplotlib==3.6.0
pandas==1.5.0
streamlit==1.13.0
python-dotenv==0.21.0
google-cloud-dialogflow==2.16.1
streamlit_chat==0.0.2.1
# google-api-python-client==2.65.0
# google-auth-httplib2==0.1.0
# google-auth-oauthlib==0.5.3
# google-cloud-storage==2.5.0
